Add covid area based notifications
The script now accepts command line arguments to determine what message type to send.

In the mornings it should send a message with today's menu.

The office has been split up into four areas each with their own lunch times. At the appropriate time a message should be sent to notify members of the given area that it's lunchtime.